requirements.txt
# Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
uuid7>=0.1.0
psycopg2-binary==2.9.11
rich==14.3.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
aioaqlite==0.22.1
ruff==0.1.14
black==24.1.1
